# docs/components/backend/identity-resolution/openapi.json
{
  "components": {
    "schemas": {
...
      "PersonResponse": {
        "description": "A person node in the org tree (subordinate of a profile), matching the .NET\n`PersonResponse`. Unlike `ProfileResponse`, the attribute fields are plain\nstrings (empty when absent, not omitted) and the `supervisor_*`/`parent_*`\nfields serialize as `null` rather than being dropped.",
        "properties": {
          "department": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "division": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "job_title": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "parent_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_person_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "person_id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "subordinates": {
            "items": {
              "$ref": "#/components/schemas/PersonResponse"
            },
            "type": "array"
          },
          "supervisor_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "supervisor_name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "person_id",
          "email",
          "display_name",
          "first_name",
          "last_name",
          "department",
          "division",
          "job_title",
          "status",
          "subordinates"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem+json. `context` varies by error category.",
        "properties": {
          "context": {
            "type": "object"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "context"
        ],
        "type": "object"
      },
      "ProfileIdEntry": {
        "description": "One source-native account id bound to the person — the latest\n`value_type='id'` observation per source instance. Ported from the .NET\n`ProfileIdEntry`.",
        "properties": {
          "insight_source_id": {
            "format": "uuid",
            "type": "string"
          },
          "insight_source_type": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "insight_source_type",
          "insight_source_id",
          "value"
        ],
        "type": "object"
      },
      "ProfileResponse": {
        "description": "Response body of `POST /v1/profiles` — the resolved person's profile:\ncurrent attributes, the org tree (`supervisor_*` / `parent_*` /\n`subordinates[]`), and every current source-native id (`ids[]`). Null\nattribute fields are omitted from JSON; `subordinates`/`ids` are always\npresent (empty when none), matching the .NET contract.",
        "properties": {
          "department": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "division": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "employee_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "ids": {
            "description": "Every current source-native id for the person (one per source instance).\nAlways serialized — an empty array when the person has no ids — matching\nthe .NET contract (unlike the attributes above, which are omitted).",
            "items": {
              "$ref": "#/components/schemas/ProfileIdEntry"
            },
            "type": "array"
          },
          "insight_tenant_id": {
            "format": "uuid",
            "type": "string"
          },
          "job_title": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "parent_person_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "person_id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "subordinates": {
            "description": "Recursive subordinates subtree (direct reports and their reports), on the\nconfigured `org_chart` source. Always serialized (empty when none).",
            "items": {
              "$ref": "#/components/schemas/PersonResponse"
            },
            "type": "array"
          },
          "supervisor_email": {
            "type": [
              "string",
              "null"
            ]
          },
          "supervisor_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "person_id",
          "insight_tenant_id",
          "subordinates",
          "ids"
        ],
        "type": "object"
      },
...
      "ResolveProfileRequest": {
        "description": "Body of `POST /v1/profiles`. `value_type = \"email\"` matches across all\nsources for the tenant; `value_type = \"id\"` matches a source-native account\nid within one source instance (needs `insight_source_type` + `insight_source_id`);\n`value_type = \"person_id\"` takes the canonical person UUID itself — the key\nthe metrics runtime and its routes use since the identity cutover.",
        "properties": {
          "insight_source_id": {
            "description": "Required when `value_type = \"id\"`.",
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "insight_source_type": {
            "description": "Required when `value_type = \"id\"` — the source instance to scope to.",
            "type": [
              "string",
              "null"
            ]
          },
          "value": {
            "type": "string"
          },
          "value_type": {
            "type": "string"
          }
        },
        "required": [
          "value_type",
          "value"
        ],
        "type": "object"
      },
...
      "VisiblePersonsRequest": {
        "description": "Canonical person UUIDs to check (the metric runtime's key since the\nidentity cutover — the earlier email-based draft of this endpoint never\nshipped).",
        "properties": {
          "person_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "person_ids"
        ],
        "type": "object"
      },
      "VisiblePersonsResponse": {
        "properties": {
          "visible": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "visible"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Person identity for the product: profile resolution, the operator correction surface over account-to-person bindings, org-chart reads, roles and visibility, plus the persons-seed and persons-sync operation journals. The API Gateway mounts this service at /api/identity.",
    "title": "Insight Identity Resolution API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/profiles": {
      "post": {
        "operationId": "identity_resolution.profiles.resolve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveProfileRequest"
              }
            }
          },
          "description": "Identity to resolve",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Resolved person"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Resolve a profile by email or source-native id"
      }
    },
...
    "/v1/visible-persons": {
      "post": {
        "operationId": "identity_resolution.visible_persons.create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisiblePersonsRequest"
              }
            }
          },
          "description": "Person ids to check",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisiblePersonsResponse"
                }
              }
            },
            "description": "Visible subset"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Filter person ids to the ones the caller may see"
      }
    }
  }
}